+++ fluentvoice/__init__.py
@@ -2,6 +2,6 @@
 Created by Nick Otmazgin
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "Nick Otmazgin"
 __license__ = "MIT"